% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_quadkey.R
\name{filter_by_quadkey}
\alias{filter_by_quadkey}
\title{Filter Performance Tiles by Quadkey}
\usage{
filter_by_quadkey(tiles, bbox)
}
\arguments{
\item{tiles}{Tiles to filter, from \code{get_performance_tiles()}}

\item{bbox}{Bounding box to pull tiles from, from \code{st_bbox()}. This way you can just get the specific tiles that you need for an analysis}
}
\value{
Filtered tiles with the same class as the \code{tiles} input, either data frame or \code{sf}
}
\description{
Filter Performance Tiles by Quadkey
}
