% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance_tiles.R
\name{get_performance_tiles}
\alias{get_performance_tiles}
\title{Pull Performance Tiles}
\usage{
get_performance_tiles(service = c("mobile", "fixed"), quarter_start,
  quarter_end, sf = FALSE)
}
\arguments{
\item{service}{Either 'mobile' or 'fixed' depending on whether you want mobile network or fixed broadband data}

\item{sf}{Default `FALSE`. Return object as `sf` dataframe}
}
\value{
either a tibble or `sf` dataframe
}
\description{
Pull Performance Tiles
}
\examples{

}
