% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooklaOpenDataR-package.R
\docType{package}
\name{ooklaOpenDataR-package}
\alias{ooklaOpenDataR}
\alias{ooklaOpenDataR-package}
\title{ooklaOpenDataR: Work with 'Speedtest by Ookla' Global Internet Performance Data}
\description{
The 'Speedtest by Ookla' Global Fixed and Mobile
    Network Performance Maps open data set provides network performance
    metrics in zoom level 16 web Mercator tiles (approximately 610.8
    meters by 610.8 meters at the equator). Download speed, upload speed,
    and latency are collected via the 'Speedtest by Ookla' applications
    for 'Android' and 'iOS' and averaged for each tile.  Measurements are
    filtered to results containing GPS-quality location accuracy.
    'ooklaOpenDataR' includes tools for retrieving and working with this
    data set.
}
\author{
\strong{Maintainer}: Katie Jolly \email{katie.jolly@ookla.com}

Authors:
\itemize{
  \item Sheryl Petersen
  \item Brian Connelly (\href{https://orcid.org/0000-0002-9948-0379}{ORCID})
}

Other contributors:
\itemize{
  \item Ookla \email{opendata@ookla.com} [copyright holder, funder]
}

}
\keyword{internal}
