% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_quadkey.R
\name{filter_by_quadkey}
\alias{filter_by_quadkey}
\title{Filter Tiles by Quadkey}
\usage{
filter_by_quadkey(tiles, bbox)
}
\arguments{
\item{tiles}{From \code{get_performance_tiles()}}

\item{bbox}{\code{\link[sf:st_bbox]{sf::st_bbox()}} bounding box describing area from which to include tiles.}
}
\value{
A filtered version of the \code{tiles} input
}
\description{
\code{filter_by_quadkey()} uses a bounding box to filter the tiles using the quadkey system as an efficient alternative to a spatial join.
}
\examples{
\dontrun{
# Filters tiles to a bounding box specified by coordinates
filter_by_quadkey(tiles, bbox = sf::st_bbox(c(xmin = 16.1, xmax = 16.6, ymax = 48.6, ymin = 47.9), crs = st_crs(4326)))

# Filters tiles to a bounding box specified by an `sf` object
nc <- st_read(system.file("gpkg/nc.gpkg", package="sf"), quiet = TRUE)
filter_by_quadkey(tiles, bbox = sf::st_bbox(nc))
}
}
