% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance_tiles.R
\name{get_performance_tiles}
\alias{get_performance_tiles}
\title{Pull Performance Tiles}
\usage{
get_performance_tiles(
  service = c("mobile", "fixed"),
  quarter,
  bbox = NULL,
  sf = FALSE,
  ...
)
}
\arguments{
\item{service}{Either 'mobile' or 'fixed' depending on whether you want mobile network or fixed broadband data}

\item{quarter}{Date indicating the start of the desired quarter (e.g., "2020-04-01" for 2020-Q2)}

\item{bbox}{Bounding box to pull tiles from. This way you can just get the specific tiles that you need for an analysis}

\item{sf}{Default \code{FALSE}. Return object as \code{sf} dataframe}

\item{...}{Additional arguments passed to \link[arrow:read_parquet]{arrow::read_parquet}}
}
\value{
either a data frame or \code{sf} data frame with the following fields:
\itemize{
\item \code{avg_d_kbps}: The average download speed of all tests performed in the tile, represented in kilobits per second
\item \code{avg_u_kbps}: The average upload speed of all tests performed in the tile, represented in kilobits per second
\item \code{avg_lat_ms}: The average latency of all tests performed in the tile, represented in milliseconds
\item \code{tests}: The number of tests taken in the tile
\item \code{devices}: The number of unique devices contributing tests in the tile
\item \code{quadkey}: The quadkey representing the tile. Quadkeys can act as a unique identifier for the tile.
}
}
\description{
\code{get_performance_tiles()} retrieves Ookla broadband performance data for a given quarter.
Data are grouped and summarized by \href{https://docs.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{quadkey} using a zoom level of 16.
This equates to a tile that is approximately 610.8 meters by 610.8 meters at the equator (18 arcsecond blocks).
}
\examples{
\dontrun{
# Pulls all fixed broadband tiles from Q2 2020
get_performance_tiles(service = "fixed", quarter = "2020-04-01", sf = TRUE)
}
}
