% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooklaOpenDataR-package.R
\docType{package}
\name{ooklaOpenDataR-package}
\alias{ooklaOpenDataR}
\alias{ooklaOpenDataR-package}
\title{ooklaOpenDataR: Speedtest by Ookla Global Fixed and Mobile Network
    Performance Map Tiles}
\description{
This dataset provides global fixed broadband and
    mobile (cellular) network performance metrics in zoom level 16 web
    mercator tiles (approximately 610.8 meters by 610.8 meters at the
    equator).  Data is provided in both Shapefile format as well as Apache
    Parquet with geometries represented in Well Known Text (WKT) projected
    in EPSG:4326.  Download speed, upload speed, and latency are collected
    via the Speedtest by Ookla applications for Android and iOS and
    averaged for each tile.  Measurements are filtered to results
    containing GPS-quality location accuracy.
}
\author{
\strong{Maintainer}: Katie Jolly \email{TODO@ookla.com}

Authors:
\itemize{
  \item Sheryl Petersen
  \item Brian Connelly (0000-0002-9948-0379)
}

Other contributors:
\itemize{
  \item Ookla \email{opendata@ookla.com} [copyright holder, funder]
}

}
\keyword{internal}
